refactor packages and hierarchy
