+++ src/main/java/com/pau_pau/project/data/models/films/Film.java
@@ -1,7 +1,9 @@
-package com.pau_pau.project.data.models;
+package com.pau_pau.project.data.models.films;
 
 import com.fasterxml.jackson.annotation.JsonIdentityInfo;
 import com.fasterxml.jackson.annotation.ObjectIdGenerators;
+import com.pau_pau.project.data.models.directors.Director;
+import com.pau_pau.project.data.models.directors.DirectorDTO;
 
 import javax.persistence.*;
 import java.util.Date;
