+++ src/main/java/com/pau_pau/project/data/models/films/FilmDTO.java
@@ -1,5 +1,7 @@
-package com.pau_pau.project.data.models;
+package com.pau_pau.project.data.models.films;
 
+import com.pau_pau.project.data.models.directors.Director;
+import com.pau_pau.project.data.models.directors.DirectorDTO;
 import io.swagger.annotations.ApiModelProperty;
 
 import java.sql.Timestamp;
